Clean up all the conditional compilation blocks and convert into conditional methods for types outside the TypeLib API
