+++ Rubberduck.Core/UI/Command/RegexSearchReplaceCommand.cs
@@ -4,9 +4,7 @@ using Rubberduck.Parsing.Common;
 
 namespace Rubberduck.UI.Command
 {
-#if !DEBUG
     [Disabled]
-#endif
     [ComVisible(false)]
     public class RegexSearchReplaceCommand : CommandBase
     {
