+++ Rubberduck.Main/Root/RubberduckIoCInstaller.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Reflection;
@@ -575,17 +576,24 @@ namespace Rubberduck.Root
 
         private Type[] RubberduckCommandBarItems()
         {
-            return new[]
+            var types = new List<Type>
             {
                 typeof(ReparseCommandMenuItem),
                 typeof(ShowParserErrorsCommandMenuItem),
                 typeof(ContextSelectionLabelMenuItem),
                 typeof(ContextDescriptionLabelMenuItem),
-                typeof(ReferenceCounterLabelMenuItem),
-#if DEBUG
-                typeof(SerializeProjectsCommandMenuItem)
-#endif
+                typeof(ReferenceCounterLabelMenuItem)
             };
+
+            AttachRubberduckDebugCommandBarItems(ref types);
+
+            return types.ToArray();
+        }
+
+        [Conditional("DEBUG")]
+        private static void AttachRubberduckDebugCommandBarItems(ref List<Type> types)
+        {
+            types.Add(typeof(SerializeProjectsCommandMenuItem));
         }
 
         private void RegisterParentMenus(IWindsorContainer container)
