+++ Rubberduck.VBEEditor/ComManagement/IComSafe.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
 
 namespace Rubberduck.VBEditor.ComManagement
@@ -7,7 +8,7 @@ namespace Rubberduck.VBEditor.ComManagement
     {
         void Add(ISafeComWrapper comWrapper);
         bool TryRemove(ISafeComWrapper comWrapper);
-#if TRACE_COM_SAFE
+
         /// <summary>
         /// Available only if the compilation constant TRACE_COM_SAFE is set. Provide a mechanism for serializing both
         /// a snapshot of the COM safe at the instant and a historical activity log
@@ -15,6 +16,5 @@ namespace Rubberduck.VBEditor.ComManagement
         /// </summary>
         /// <param name="targetDirectory">The path to a directory to place the serialized files in</param>
         void Serialize(string targetDirectory);
-#endif
     }
 }
