+++ Rubberduck.VBEEditor/ComManagement/StrongComSafe.cs
@@ -1,10 +1,7 @@
 ﻿using System.Collections.Concurrent;
-using Rubberduck.VBEditor.SafeComWrappers.Abstract;
-
-#if TRACE_COM_SAFE
-using System.Linq;
 using System.Collections.Generic;
-#endif
+using System.Linq;
+using Rubberduck.VBEditor.SafeComWrappers.Abstract;
 
 namespace Rubberduck.VBEditor.ComManagement
 {
@@ -22,16 +19,12 @@ namespace Rubberduck.VBEditor.ComManagement
                     comWrapper, 
                     key =>
                     {
-#if TRACE_COM_SAFE
                         TraceAdd(comWrapper);
-#endif
                         return 1;
                     }, 
                     (key, value) =>
                     {
-#if TRACE_COM_SAFE
                         TraceUpdate(comWrapper);
-#endif
                         return value;
                     });
             }
@@ -45,9 +38,8 @@ namespace Rubberduck.VBEditor.ComManagement
             }
 
             var result = _comWrapperCache.TryRemove(comWrapper, out _);
-#if TRACE_COM_SAFE
             TraceRemove(comWrapper, result);
-#endif
+
             return result;
         }
 
@@ -70,12 +62,10 @@ namespace Rubberduck.VBEditor.ComManagement
             _comWrapperCache.Clear();
         }
 
-#if TRACE_COM_SAFE
         protected override IDictionary<int, ISafeComWrapper> GetWrappers()
         {
             return _comWrapperCache.Keys.ToDictionary(GetComWrapperObjectHashCode);
         }
-#endif
     }
 }
