+++ Rubberduck.VBEEditor/ComManagement/WeakComSafe.cs
@@ -1,11 +1,8 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
-
-#if TRACE_COM_SAFE
 using System.Linq;
 using System.Collections.Generic;
-#endif
 
 namespace Rubberduck.VBEditor.ComManagement
 {
@@ -22,16 +19,12 @@ namespace Rubberduck.VBEditor.ComManagement
                     GetComWrapperObjectHashCode(comWrapper), 
                     key =>
                     {
-#if TRACE_COM_SAFE
                         TraceAdd(comWrapper);
-#endif
                         return (DateTime.UtcNow, new WeakReference<ISafeComWrapper>(comWrapper));
                     },
                     (key, value) =>
                     {
-#if TRACE_COM_SAFE
                         TraceUpdate(comWrapper);
-#endif
                         return (value.insertTime, new WeakReference<ISafeComWrapper>(comWrapper));
                     });
             }
@@ -46,9 +39,8 @@ namespace Rubberduck.VBEditor.ComManagement
             }
 
             var result = _comWrapperCache.TryRemove(GetComWrapperObjectHashCode(comWrapper), out _);
-#if TRACE_COM_SAFE
             TraceRemove(comWrapper, result);
-#endif
+
             return result;
         }
 
@@ -74,7 +66,6 @@ namespace Rubberduck.VBEditor.ComManagement
             _comWrapperCache.Clear();
         }
 
-#if TRACE_COM_SAFE
         protected override IDictionary<int, ISafeComWrapper> GetWrappers()
         {
             var dictionary = new Dictionary<int, ISafeComWrapper>();
@@ -89,7 +80,6 @@ namespace Rubberduck.VBEditor.ComManagement
 
             return dictionary;
         }
-#endif
     }
 }
