+++ Rubberduck.VBEEditor/WindowsApi/WM.cs
@@ -942,12 +942,8 @@ namespace Rubberduck.VBEditor.WindowsApi
         RUBBERDUCK_SINKING = USER + 0x0D1E,
     }
 
-#if (DEBUG && (THIRSTY_DUCK || THIRSTY_DUCK_WM))
-
     public static class WM_MAP
     {
         public static readonly Dictionary<uint, string> Lookup = EnumHelper.ToDictionary<WM, uint>();
     }
-
-#endif
 }
